+++ tests/urls.py
@@ -1,10 +1,11 @@
 from django.contrib import admin
-from django.urls import path
+from django.urls import include, path
 
 from tests.app_teste import views
 
 urlpatterns = [
     path("gestao-interna/", admin.site.urls),
+    path("", include("infra_vibecoding.login.urls")),
     path("", views.inicio),
     path("painel/", views.painel),
     path("pedidos/novo/", views.novo_pedido),